* sysdeps/i386/fpu/ftestexcept.c (fetestexcept): Use short for fnstsw.

+++ sysdeps/i386/fpu/ftestexcept.c
@@ -26,7 +26,7 @@
 int
 fetestexcept (int excepts)
 {
-  int temp;
+  short temp;
   int xtemp = 0;
 
   /* Get current exceptions.  */
